tugasku-backend/requirements.txt
pyramid>=2.0.2
pyramid_jinja2>=2.10
waitress>=2.1.2
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9
bcrypt>=4.1.2
PyJWT>=2.8.0
alembic>=1.13.1
pytest>=7.4.3
pytest-cov>=4.1.0
webtest>=3.0.0
